requirements.txt
alembic==1.16.4
anyio==4.10.0
asyncio==4.0.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
passlib==1.7.4
psycopg2-binary==2.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.3.2
numpy==1.26.4
pandas==2.1.4
pymongo==4.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.4
resend==2.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==8.5.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
Werkzeug==3.1.3
transformers
torch
sentence_transformers
faiss-cpu
flask[async]
asgiref
pydantic
nest_asyncio
google-genai
google-generativeai